pyproject.toml
[tool.poetry]
name = "skqulacs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
numpy = "^1.21.4"
scipy = "^1.7.2"
scikit-learn = "^1.0.1"
qulacs-osaka = {git = "https://github.com/Qulacs-Osaka/qulacs-osaka", rev = "dev"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
matplotlib = "^3.5.0"
pandas = "^1.3.4"
black = "^21.11b1"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.910"
Sphinx = "^4.3.0"
sphinx-book-theme = "^0.1.7"
ipykernel = "^6.5.0"
myst-nb = "^0.13.1"

[tool.pysen]
version = "0.9"

[tool.isort]
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
